All servers use event based framework now. Only thing left unimplemented is sending entity updates from Map server.
git-svn-id: svn://svn.code.sf.net/p/segs/code2@178 a19be42f-918d-4297-8077-ab25d1e4f51f

+++ Projects/CoX/Common/Servers/src/Database.cpp
@@ -171,7 +171,7 @@ size_t DbResultRow::getColFloatArray(const char *column_name,float *arr,size_t a
 	while(res && idx<arr_size)
 	{
 		res++;
-		arr[idx++]=atof(res);
+		arr[idx++]=(float)atof(res);
 		res=strchr(res,',');
 	}
 	return idx;
